Short release notes for app version 0.5.5.24:
0.5.5.24
#46 User account email format validation rules not defined
changed regex to
/^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,4}$/i


0.5.5.23
added possibility to handle the error message inside the viewModel (or failCallback) and avoid the default error message.
Currenty implemented:
For example for scheduler, if the user has no scheduler yet he gets an error message "No schedule data found for this robot". But 
inside viewModel we handled this error and created an empty scheduler. So the error message was not realy relevant for the user as he 
doesn't had any negative effects.
New solution:
There is a new argument for failCallbacks: function(error, notificationOptions, errorHandled), where errorHandled is a deferred object
in the communication wrapper onCallbackReturn method.    
If the viewModel handles the error it must call errorHandled.resolve() now, to inform the errorHandledDone promise object that the 
error has been handled and the default error message doesn't need to be shown.


0.5.5.22
#81 Two cleaning events on one day
fixed wrong weekindex for german language for "create new schedule event on tap"
added check for "add" schedule button if there is a free day left 


0.5.5.21
#103 menu won't open with the samsung tablet
added translateZ(0) to dialog

0.5.5.20
#105 possibility to slide the menu / screen view 
added correct high res values
added overflow hidden to suppress scrolling

0.5.5.19
#98 Link to Vorwerk Homepages
                             
0.5.5.18
#102 Send to Base message missing on small display

0.5.5.17
#101 Misspelling of "address" in User Settings

0.5.5.16
#99 Unexpected New cleaning event
using click instead of vclick to prevent that the event is triggered to the underlaying page

0.5.5.15
#80 New scheduler event suggests cleaning time 00:00
fixed create a new event on tap on scheduler (on some devices the event.offsetY was not available)