Android SmartApp and Simulator: Support for Robot Linking
Initial check-in to add "Robot Linking" support to associate a
robot with a specific user.

- Robot first calls an API to generate the link code. Server returns 
  the link code and the time until which the link code is valid.
- Robot shows the link code in its UI to the user and starts a timer 
  for the link code expiry
- User then enters the link code on SmartApp side
- This link code is validated by server and notification is sent 
  to the user along with the email id of the user
- Robot displays the Notification to the user to accept/reject the linking 
- If user accept the linking, robot gets associated with the user 

APIs for Phone Gap UI:
- Added API initiate linking, which takes linking code and user's email id

NOTE: Currently the backend functionality to enable this support is ONLY
available on NeatoDev. We haven't pushed the changes to Neato Staging since
there is a user testing exercise in China on 09/18/2013. So, if you want
to test this feature, you will need to change the code to build both the 
SmartApp and the Simulator to point to NeatoDev.