iOS SmartApp Enhancements
Issue #340 New Feature: App can expire
added expire message with an exit button ("beenden") which closes the app.
IMPORTANT: this is only for the beta refresh, afterward this modification needs to be removed. Otherwise the app won't be accepted by Apple for the AppStore according it's not allowed to close an app.

The iOS Human Interface Guidelines specify,

"Always Be Prepared to Stop iOS applications stop when people press the Home button to open a different application or use a device feature, such as the phone. In particular, people don’t tap an application close button or select Quit from a menu. To provide a good stopping experience, an iOS application should:

Save user data as soon as possible and as often as reasonable because an exit or terminate notification can arrive at any time.

Save the current state when stopping, at the finest level of detail possible so that people don’t lose their context when they start the application again. For example, if your app displays scrolling data, save the current scroll position."

> It would be appropriate to remove any mechanisms for quitting the app.